Move TestImportGenesetData back to own class
